pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "apscheduler>=3.11.0",
    "asyncio-mqtt>=0.16.2",
    "asyncpg>=0.30.0",
    "celery>=5.5.3",
    "cryptography>=45.0.4",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "python-telegram-bot==20.7",
    "qrcode>=8.2",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "telegram>=0.0.1",
    "trafilatura>=2.0.0",
    "uvloop>=0.21.0",
    "werkzeug>=3.1.3",
]
